Healthcheck

+++ handler.go
@@ -7,15 +7,20 @@ import (
 	"github.com/nzgogo/micro/context"
 )
 
+var(
+	REQUEST = "request"
+	RESPONSE = "response"
+	HEALTHCHECK = "healthCheck"
+)
+
 func (s *service) ServerHandler(nMsg *nats.Msg) {
 	//decode message
 	message := &codec.Message{}
 	codec.Unmarshal(nMsg.Data, message)
-
 	sub := s.opts.Transport.Options().Subject
 
 	//check message type, response or request
-	if message.Type == "request" {
+	if message.Type == REQUEST{
 		//check if the message is a Request or Publish.
 		if nMsg.Reply != "" {
 			message.ReplyTo = nMsg.Reply
@@ -51,8 +56,22 @@ func (s *service) ServerHandler(nMsg *nats.Msg) {
 				)
 			}
 		}()
-	} else {
-		rpl := s.opts.Context.Get(message.ContextID).Request
+	} else if message.Type == HEALTHCHECK{
+		go func() {
+			checkStatus, feedback := healthCheck(s.config)
+			msg := codec.NewResponse("",checkStatus, feedback, nil)
+			replyBody,_ :=codec.Marshal(msg)
+			s.opts.Transport.Publish(nMsg.Reply,replyBody)
+
+		}()
+
+	} else if message.Type == RESPONSE {
+		conversation := s.opts.Context.Get(message.ContextID)
+		if conversation == nil {
+			return
+		}
+
+		rpl := conversation.Request
 		s.opts.Transport.Publish(rpl, nMsg.Data)
 		s.opts.Context.Delete(message.ContextID)
 	}
@@ -63,16 +82,28 @@ func (s *service) ApiHandler(nMsg *nats.Msg) {
 	message := &codec.Message{}
 	codec.Unmarshal(nMsg.Data, message)
 	ctx := s.opts.Context
+	if message.Type == HEALTHCHECK{
+		go func() {
+			checkStatus, feedback := healthCheck(s.config)
+			msg := codec.NewResponse("",checkStatus, feedback, nil)
+			replyBody,_ :=codec.Marshal(msg)
+			s.opts.Transport.Publish(nMsg.Reply,replyBody)
 
-	r := ctx.Get(message.ContextID).Response
+		}()
 
-	fn := gogoapi.WriteResponse
-	for i := len(s.opts.HttpRespWrappers); i > 0; i-- {
-		fn = s.opts.HttpRespWrappers[i-1](fn)
+	} else if message.Type == RESPONSE {
+		conversation := ctx.Get(message.ContextID)
+		if conversation == nil {
+			return
+		}
+		r := conversation.Response
+
+		fn := gogoapi.WriteResponse
+		for i := len(s.opts.HttpRespWrappers); i > 0; i-- {
+			fn = s.opts.HttpRespWrappers[i-1](fn)
+		}
+		fn(r, message)
+		ctx.Done(message.ContextID)
+		ctx.Delete(message.ContextID)
 	}
-	fn(r, message)
-
-	ctx.Done(message.ContextID)
-	ctx.Delete(message.ContextID)
-
 }
